@@ -1,28 +1,9 @@
 package com.lavishly.android.SubsidiaryActivities;
 
-import android.annotation.SuppressLint;
-import android.app.Dialog;
-import android.content.Context;
-import android.content.SharedPreferences;
-import android.media.Image;
-import android.support.v7.app.AppCompatActivity;
+import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
-import android.view.LayoutInflater;
-import android.view.MenuItem;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.AdapterView;
-import android.widget.BaseAdapter;
-import android.widget.Button;
-import android.widget.ImageView;
-import android.widget.ListView;
-import android.widget.TextView;
 
-import com.facebook.shimmer.ShimmerFrameLayout;
 import com.lavishly.android.R;
-
-import static com.lavishly.android.R.id.diaBoxPersonage;
-import static com.lavishly.android.R.id.display_always;
 
 public class PersonageActivity extends AppCompatActivity {
 
